Move to formatter that accepts data argument
